Add interactivity to the list
This is a significant part of my jurney so far
I have been waiting for the day where i'd be able to create a todo list
even is it will be a miniature one. and i did it today
i am really happy - i also created a rock paper scissors game few daya
back.